+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Feat: update coordinator API to show more details on pending channel balance
 - Feat: show dlc-channel balance instead of ln-balance in app and in coordinator's API
 - Chore: don't allow multiple dlc-channels per user
+- Feat: show dlc-channel opening transaction in transaction history
 
 ## [1.7.4] - 2023-12-20